pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "kora-mcp-server"
version = "1.3.0"
description = "MCP server for Kora — deterministic authorization for AI agent spending"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
  {name = "Idkasam", email = ""}
]
keywords = ["mcp", "kora", "authorization", "ai-agent", "payments"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Office/Business :: Financial",
]
dependencies = [
  "mcp>=1.0.0",
  "httpx>=0.25.0",
  "PyNaCl>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/Idkasam/kora-mcp-server"
Repository = "https://github.com/Idkasam/kora-mcp-server"
Documentation = "https://github.com/Idkasam/Kora"

[project.scripts]
kora-mcp = "kora_mcp.server:main"
